Player Character and its Potentials
Almost done with the player character and it is giving me quite a bit to reflect on how to expand new potentials out of it. Eerily, it feels similar to the beginning of my master's thesis, developing the introspective rulebook method when every action I did in the game, was a possible point of inquiry for experimentation. This can become really confusing and overwhelming fast but I would like to experience it first before trying to come up with ways to deal with it.