Shooting
Implemented shooting. Any modifications I could make to it, the changes feel the most traditionally conforming ones. In most games that contain shooting, there is also differences in the way you can shoot. It creates a diversity. So in this one it might make sense to keep 'shooting' as one simple thing for now. Or explore the space it creates. One of the common themes that emerge out of this way of game making is that there are some commonalities between the things you can change in different things. The size, the quantity; if it is physical; the weight, friction, etc.. Right now I am just focused on finishing this 'closed game', and it is giving me plenty of leads to explore.