Added The Second player + Tidbit Thoughts on Player Agency on Rules
1. For some rules, I can give them to you deconstructed, and let you decide how do you want to set them up.
2. For others, I can set them up as I like and let you decide whether you want them or not.
Example:
With the inputs, either I can let you decide what inputs are what, or I can create a new option 'randomized inputs', that takes agency away from you but lets you experience something unknown.

This 'othering' effect of games is important, I will be on the lookout for it when I start working on the project.