Some Fixes, Rearrangements and Thoughts
I am starting to feel I am inching closer to the first iteration of this game but an important area needs to be designed first. How do you communicate to the players that the more they can get creative with the rules, the more they will enjoy this game? I am trying to prompt them in the internal rulebook, but I believe the aesthetic and the theming of the game should also directly communicate this.