Implemented Gravity Change
Next urgent thing I need to do is implement the game states loop so I can have something playable and then start polishing. I am getting close to exhausting the things I can open. So after that, I need to figure out how to also add new stuff and open it too. I think that would be an interesting phase to see how the new things and the existing things interact with each other as they are being designed in an open manner.