Implemented Round Progression and Scores and Basic UI Theming
Quite a lot on this one. I think the biggest one was the revelation that for some rules, translating them from a material rule to an internal rule opens up possibilities, and can help me avoid some extra work. I will be on the lookout for other applications of how I might translate rules.