Implemented Input Changes
This was the toughest one so far. Not because it was complex on its own that much, but because there are so many inputs that need to be handled manually. I am really glad I implemented this one however. I think the inputs/control schemes is something we very rarely think about in the context of digital games and it feels good to make it accesible to player to play with this aspect of games. And i think with this one I am officially done with the features I wanted to implemented for the first iteration.