Added Options to Change Blocking Parameters
In a playtest one of the players complained how easy it was to just exploit the blocking. Now I am giving them the option to change the parameters relating to it. It is interesting how putting one of these mechanics in an unchangeable manner immediately stood out to them. They wanted to gain control to it.

Next I want to implement tooltips and also an introduction text to the game.
After that I will implement dashing.