Started implementing dash
I think we have a good amount of features, despite no weapon variation. Which is maybe fine? Maybe not? Weapons are a huge area of expansion so I think that is why I feel some resistance to start making things on that. Anyway, soon I will add the dash variables to the UI. This one will be a bit tricky because a dash has more variables than you would think (or at least the way i implemented it), but giving access to all of this will muddy the waters in my opinion. So I will instead choose the ones that really matter. My expectation is that it will also have quite a bit of game breaking ends in its parameters.